fix: Change ingredient quantity field type to float

+++ app/src/main/java/com/mattwiduch/bakeit/model/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient {
 
   @SerializedName("quantity")
   @Expose
-  private Integer quantity;
+  private Float quantity;
   @SerializedName("measure")
   @Expose
   private String measure;
@@ -19,11 +19,11 @@ public class Ingredient {
   @Expose
   private String ingredient;
 
-  public Integer getQuantity() {
+  public Float getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(Integer quantity) {
+  public void setQuantity(Float quantity) {
     this.quantity = quantity;
   }
